% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompEllipse.R
\name{CompEllipse}
\alias{CompEllipse}
\title{Uncertainty Elipses Computations}
\usage{
CompEllipse(x, xsd, y, ysd)
}
\arguments{
\item{x}{vector of x values}

\item{xsd}{vector of xsd values (specifies the width of the elipse)}

\item{y}{vector of y values}

\item{ysd}{vector of ysd values (specifies the height of the elipse)}
}
\value{
dataframe of Ellipse x-y points
}
\description{
Computes the datapoints in order to plot uncertainty elipses \cr
}
\examples{
Data <- data.frame(x=seq(1:10), xsd=runif(10), y=rnorm(n=10), ysd=runif(10))
Ellipse <- CompEllipse(x=Data$x, xsd=Data$xsd, y=Data$y, ysd=Data$ysd)
p <- ggplot()+
    geom_point(data=Data, aes(x=x, y=y))+
    geom_polygon(data=Ellipse,aes(x=xEll,y=yEll, group=obs), alpha=.15)
print(p)
}
\keyword{plotting}
