% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotRTheme.R
\name{DotRTheme}
\alias{DotRTheme}
\title{Data on the Rocks Theme}
\usage{
DotRTheme(axis.text.size = 16, axis.title.size = 16, title.size = 20,
  legend.position = "none")
}
\arguments{
\item{axis.text.size}{axis text size (defaults to 16)}

\item{axis.title.size}{axis title size (defaults to 16)}

\item{title.size}{title size (degaults to 20)}

\item{legend.position}{positions of the legend (defaults to none)}
}
\description{
Default ggplot theme for Data on the Rocks blog posts
}
\examples{
Data <- data.frame(x=seq(1:10), y=rnorm(n=10))
p <- ggplot(Data, aes(x=x, y=y))+ geom_point() +
    DotRTheme()
print(p)
}
\keyword{plotting}
