% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfAdj.R
\name{InfAdj}
\alias{InfAdj}
\title{Adjusting for Inflation}
\usage{
InfAdj(yr, adj.yr, value)
}
\arguments{
\item{yr}{Year raw}

\item{adj.yr}{Year to adjust to (e.g. in 2016 dollars)}

\item{value}{dollar vallues}
}
\value{
dataframe adding adjustment factors and inflation adjusted values
}
\description{
Function to adjust a monetary value for inflation\cr
Works for scalar and vectors
}
\examples{
InfAdj(yr=1990, adj.yr=2013, value=1)
Dollars <- data.frame(yr=seq(1990,2010), value=1)
InfAdj(yr=Dollars$yr, adj.yr=2013, value=Dollars$value)
}
\keyword{inflation}
