% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Printing Multiple Plots in Grid}
\arguments{
\item{...}{Series of plots}

\item{cols}{number of columns (defaults to 1)}

\item{layout}{specifiable layout}
}
\description{
Computes the datapoints in order to plot uncertainty elipses \cr
}
\examples{
Data <- data.frame(x1=seq(1:10), y1=runif(10), x2=rnorm(n=10), y2=runif(10))
p1 <- ggplot(Data, aes(x=x1, y=y1)) + geom_point() + SpTheme()
p2 <- ggplot(Data, aes(x=x2, y=y2)) + geom_point() + SpTheme()
multiplot(p1, p2, cols=2)
}
\keyword{plotting}
