% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCor.R
\name{SimCor}
\alias{SimCor}
\title{Simulating Correlated Data}
\usage{
SimCor(n, xmean, xsd, ymean, ysd, rho)
}
\arguments{
\item{n}{number of datapoints to simulate}

\item{xmean}{mean of x variable}

\item{xsd}{standard deviation of x variable}

\item{ymean}{mean of y variable}

\item{ysd}{standard deviation of y variable}

\item{rho}{intended correlation}
}
\value{
dataframe of correlated x and y variables
}
\description{
Simulates bivariate correlation of specified n, means, sd, and rho
}
\examples{
SimCor(n=20, xmean=10, xsd=5, ymean=1, ysd=.5, rho=.5)
}
\keyword{simulation}
