% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCorX.R
\name{SimCorX}
\alias{SimCorX}
\title{Simulating Correlated Y for Given X}
\usage{
SimCorX(x, ymean, ysd, rho)
}
\arguments{
\item{ymean}{mean of y variable}

\item{ysd}{standard deviation of y variable}

\item{rho}{intended correlation}
}
\value{
dataframe of correlated x and y variables
}
\description{
Simulates bivariate correlation of specified n, means, sd, and rho
}
\examples{
SimCorX(x, Ymean=1, Ysd=.5, rho=.5)
}
\keyword{simulation}
