% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpDesc.R
\name{SpDesc}
\alias{SpDesc}
\title{Spencer Descriptives}
\arguments{
\item{data}{dataframe to get descriptives/frequencies}
}
\description{
Computes summary descriptives and frequency tables. \cr
Uses pastecs stat.desc for numeric data. \cr
Prints tables for factor data. \cr
Able to take a vector or dataframe.
}
\examples{
SpDesc(XYData)
SpDesc(XYData$X)
}
\keyword{descriptives}
