% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpHist.R
\name{SpHist}
\alias{SpHist}
\title{Histogram with ggplot}
\arguments{
\item{data}{dataframe or vector of numeric values}

\item{variable}{string of the variable of interest}

\item{bins}{number of bins (defaults to 30)}

\item{save}{boolean of whether to save image}
}
\description{
Generate Histogram using ggplot2 with overlayed normal distribution. \cr
Also includes mean and SD in the title.
}
\examples{
SpHist(data=rnorm(100), variable="X Variable", bins=30, save=F)
}
\keyword{descriptives}
\keyword{plotting}
