% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpQQPlot.R
\name{SpQQPlot}
\alias{SpQQPlot}
\title{Q-Q Plot with ggplot}
\arguments{
\item{data}{vector of numeric values}

\item{variable}{string of the variable of interest for title}

\item{save}{boolean of whether to save image}
}
\description{
Generate Q-Q plot using ggplot2 with normal distribution reference. \cr
}
\examples{
SpQQPlot(rnorm(50), varname="Variable #1", save=T)
}
\keyword{descriptives}
\keyword{plotting}
