% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpTheme.R
\name{SpTheme}
\alias{SpTheme}
\title{Spencer's Default Theme}
\usage{
SpTheme(axis.text.size = 12, axis.title.size = 12, title.size = 16,
  legend.position = "none")
}
\arguments{
\item{axis.text.size}{axis text size (defaults to 12)}

\item{axis.title.size}{axis title size (defaults to 12)}

\item{title.size}{title size (degaults to 16)}

\item{legend.position}{positions of the legend (defaults to none)}
}
\description{
Default ggplot theme for Spencer's publications
}
\examples{
Data <- data.frame(x=seq(1:10), y=rnorm(n=10))
p <- ggplot(Data, aes(x=x, y=y))+ geom_point() +
    SpTheme()
print(p)

#Spencer's default ggplot2 theme
}
\keyword{plotting}
