% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpDesc.R
\name{SpDesc}
\alias{SpDesc}
\title{Spencer Descriptives}
\arguments{
\item{data}{dataframe to get descriptives/frequencies}
}
\description{
Computes summary descriptives and frequency tables. \cr
Uses pastecs stat.desc for numeric data. \cr
Prints tables for factor data. \cr
Able to take a vector or dataframe.
}
\examples{
Data <- data.frame(x=seq(1:10), y=rnorm(n=10), c=c(rep("A",7), rep("B",3)))
SpDesc(Data)
SpDesc(Data$x)
SpDesc(Data$c)
}
\keyword{descriptives}
